# feat: add performance metrics commands