--- a/tutoraspects/commands_v1.py
+++ b/tutoraspects/commands_v1.py
@@ -129,6 +129,7 @@ def import_assets() -> list[tuple[str, str]]:
         ),
     ]
 
+
 # Ex: "tutor local do performance-metrics "
 @click.command(context_settings={"ignore_unknown_options": True})
 def performance_metrics() -> list[tuple[str, str]]: