--- a/tutoraspects/templates/aspects/apps/superset/pythonpath/performance_metrics.py
+++ b/tutoraspects/templates/aspects/apps/superset/pythonpath/performance_metrics.py
@@ -1,0 +1,139 @@
+from superset.app import create_app
+app = create_app()
+app.app_context().push()
+
+"""
+The plan
+
+1. Load the instructor dashboard
+2. Load the instructor dashboard datasets
+3. Perform the following operations for every dataset:
+    a. Load the dataset
+    b. Perform queries on the dataset with/without filters
+    c. Perform aggregations on the dataset
+    d. Measure the time taken for each operation
+    e. Repeat 5 times and take metrics such as average, median, max, min, etc.
+"""
+from superset.extensions import db
+from flask import g
+from superset.models.dashboard import Dashboard
+from superset.models.slice import Slice
+from superset.charts.schemas import ChartDataQueryContextSchema
+from superset.charts.data.commands.get_data_command import ChartDataCommand
+from superset import security_manager
+import json
+from datetime import datetime
+import sqlparse
+from unittest.mock import patch
+import uuid
+import time
+
+ASPECTS_VERSION = "{{ASPECTS_VERSION}}"
+UUID = str(uuid.uuid4())[0:6]
+RUN_ID = f'aspects-{ASPECTS_VERSION}-{UUID}'
+
+report_format = """
+    {i}. {slice}
+    Superset time: {superset_time}"""
+
+query_format = """    Query duration: {query_duration_ms} s.
+    Result rows: {result_rows}
+    Memory Usage (MB): {memory_usage_mb}
+    Row count Clickhouse {rowcount:}
+    Filters: {filters}"""
+
+
+def performance_metrics():
+    """Measure the performance of the dashboard."""
+    with patch('clickhouse_connect.common.build_client_name') as mock_build_client_name:
+        mock_build_client_name.return_value = RUN_ID
+        embedable_dashboards = {{ SUPERSET_EMBEDDABLE_DASHBOARDS }}
+        dashboards = db.session.query(Dashboard).filter(Dashboard.slug.in_(embedable_dashboards)).all()
+        report = []
+        for dashboard in dashboards:
+            print("Dashboard:", dashboard.slug)
+            for slice in dashboard.slices:
+                report.append(meassure_chart(slice))
+        return report
+
+
+def meassure_chart(slice, extra_filters=[]):
+    """
+    Meassure the performance of a chart and return the results.
+    """
+    print("Fetching slice data:", slice)
+    query_context = json.loads(slice.query_context)
+    query_context.update({
+        "result_format": "json",
+        "result_type": "full",
+        "force": True,
+    })
+
+    if extra_filters:
+        query_context["filters"].extend(extra_filters)
+
+    g.user = security_manager.find_user(username="{{SUPERSET_ADMIN_USERNAME}}")
+    query_context = ChartDataQueryContextSchema().load(query_context)
+    command = ChartDataCommand(query_context)
+
+    start_time = datetime.now()
+    result = command.run()
+    end_time = datetime.now()
+
+    result["time_elapsed"] = '{} s.'.format((end_time-start_time).total_seconds())
+    result["slice"] = slice
+    return result
+
+def get_query_log_from_clickhouse(report):
+    """
+    Get the query log from clickhouse and print the results.
+    """
+    chart_uuid = "bb13bb31-c797-4ed3-a7f9-7825cc6dc482"
+
+    slice = db.session.query(Slice).filter(Slice.uuid == chart_uuid).one()
+
+    query_context = json.loads(slice.query_context)
+    query_context["queries"][0]["filters"].append(
+        {
+            'col': 'http_user_agent', 'op': '==', 'val': RUN_ID
+        }
+    )
+    slice.query_context = json.dumps(query_context)
+
+    result = meassure_chart(slice)
+
+    clickhouse_queries = {}
+    for query in result["queries"]:
+        for row in query["data"]:
+            parsed_sql = str(sqlparse.parse(row.pop("query"))[0])
+            clickhouse_queries[parsed_sql] = row
+
+
+    print(f"\nSuperset Reports: {RUN_ID}")
+    for i, result in enumerate(report):
+        print(report_format.format(
+            i=(i+1),
+            slice=result["slice"],
+            superset_time=result["time_elapsed"]
+        ))
+        for i, query in enumerate(result["queries"]):
+            parsed_sql = str(sqlparse.parse(query["query"])[0]).replace(";", "") + "\n FORMAT Native"
+            clickhouse_report = clickhouse_queries.get(parsed_sql, {})
+            print(query_format.format(
+                query_duration_ms=clickhouse_report.get("query_duration_ms") / 1000,
+                memory_usage_mb =clickhouse_report.get("memory_usage_mb"),
+                result_rows=clickhouse_report.get("result_rows"),
+                rowcount=query["rowcount"],
+                filters=query["applied_filters"],
+            ))
+
+
+
+
+if __name__ == "__main__":
+    print("Running performance metrics. RUN ID:", RUN_ID)
+    report = performance_metrics()
+    # Clickhouse query log takes some seconds to log queries.
+    print("Waiting for clickhouse log...")
+    time.sleep(10)
+    get_query_log_from_clickhouse(report)