--- a/tutoraspects/templates/openedx-assets/queries/fact_course_grades.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_course_grades.sql
@@ -3,7 +3,8 @@ with grades as (
     from {{ DBT_PROFILE_TARGET_DATABASE }}.fact_grades
     where grade_type = 'course'
     {% raw %}
-    {% if get_filters('course_name', remove_filter=True) != [] and filter_values('course_name') != [] %}
+    {% if get_filters('course_name', remove_filter=True) == [] %}
+    {% elif filter_values('course_name') != [] %}
     and entity_name in {{ filter_values('course_name', remove_filter=True) | where_in }}
     {% else %}
     and 1=0