--- a/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_course_summary.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_course_summary.sql
@@ -134,7 +134,7 @@ SELECT
 FROM summary
 WHERE
     {% raw %}
-    {% if get_filters('course_name', remove_filter=True) != [] && filter_values('course_name') != [] %}
+    {% if get_filters('course_name', remove_filter=True) != [] and filter_values('course_name') != [] %}
     course_name in {{ filter_values('course_name') | where_in }}
     {% else %}
     1=0