--- a/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_course_summary.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_course_summary.sql
@@ -135,6 +135,7 @@ FROM summary
 WHERE
     {% raw %}
     {% if get_filters('course_name', remove_filter=True) == [] %}
+    1=1
     {% elif filter_values('course_name') != [] %}
     course_name in {{ filter_values('course_name') | where_in }}
     {% else %}