--- a/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_summary.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_summary.sql
@@ -118,7 +118,7 @@ SELECT
 FROM summary
 where
     {% raw %}
-    {% if get_filters('problem_name_with_location', remove_filter=True) != [] && filter_values('problem_name_with_location') != [] %}
+    {% if get_filters('problem_name_with_location', remove_filter=True) != [] and filter_values('problem_name_with_location') != [] %}
     problem_name_with_location in {{ filter_values('problem_name_with_location') | where_in }}
     {% else %}
     1=0