--- a/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_summary.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_learner_problem_summary.sql
@@ -119,6 +119,7 @@ FROM summary
 where
     {% raw %}
     {% if get_filters('problem_name_with_location', remove_filter=True) == [] %}
+    1=1
     {% elif filter_values('problem_name_with_location') != [] %}
     problem_name_with_location in {{ filter_values('problem_name_with_location') | where_in }}
     {% else %}