--- a/tutoraspects/templates/openedx-assets/queries/fact_problem_grades.sql
+++ b/tutoraspects/templates/openedx-assets/queries/fact_problem_grades.sql
@@ -5,7 +5,8 @@ with grades as (
         grade_type = 'problem'
 
         {% raw %}
-        {% if get_filters('problem_name_with_location', remove_filter=True) != [] and filter_values('problem_name_with_location') != [] %}
+        {% if get_filters('problem_name_with_location', remove_filter=True) != [] %}
+        {% elif filter_values('problem_name_with_location') != [] %}
         and entity_name_with_location in {{ filter_values('problem_name_with_location', remove_filter=True) | where_in }}
         {% else %}
         and 1=0