### Description

This PR does the following changes:

1. Restore the `query_context` param in charts to query data from charts programmatically.
2. Fixes the datasets's Jinja template so that the data is only filtered on the dashboard context where the filters are defined.
3. Adds a `performance-metrics` command that queries the data on all the `SUPERSET_EMBEDDABLE_DASHBOARDS` charts and gets the data needed for performance measurements. Every query performed (by clickhouse-connect) is attached a custom `http_user_agent` following this format `aspects-{ASPECTS_VERSION}-{UUID[0:6]}`  so that those can the queries can be retrieved and matched later.

An example output is here for reference:

Superset Reports: aspects-0.86.1-b327bf

```
    1. Course Grade Distribution
    Superset time: 0.906486 s.
    Query duration: 0.464 s.
    Result rows: 1
    Memory Usage (MB): 191.3544635772705
    Row count Clickhouse 1
    Filters: [{'column': 'emission_time'}]

    2. Distribution Of Hints Per Correct Answer
    Superset time: 2.238558 s.
    Query duration: 1.469 s.
    Result rows: 0
    Memory Usage (MB): 377.11946868896484
    Row count Clickhouse 0
    Filters: []

    3. Posts per user
    Superset time: 0.131786 s.
    Query duration: 0.007 s.
    Result rows: 0
    Memory Usage (MB): 0.351043701171875
    Row count Clickhouse 0
    Filters: []
...
```

The following chart and its dataset are created to query the queries performance:

![image](https://github.com/openedx/tutor-contrib-aspects/assets/33465240/7b83356f-c07f-4b61-b9de-b6606670a833)